feat!: use versioned keys for dynamically adding and remove commit stores (#3320)
As was discovered with the apphash mismatch in
https://github.com/celestiaorg/celestia-app/issues/3167, by adding both
v1 and v2 stores in the constructor we actually end up with a different
app hash to `v1.x` as now there are a bunch of empy but initialized
stores. What's needed is a mechanism that can start with only the v1
stores and then during the upgrade process add or remove the store for
v2 as necessary.

The main problem with this is that because we are using a branch of
state we can't actually modify the stores until we write the new state
to disk. Hence we can no longer perform the migration in `EndBlock` but
in `Commit` whereby we first write state, then create the new stores,
then perform the module migrations (like call `InitGenesis`) and then
write it again finally returning the app hash.

This PR depends on https://github.com/celestiaorg/cosmos-sdk/pull/387
which introduces the upgrade hooks during `Commit`. This PR itself adds
the two hooks for updating the stores and performing the migrations.

---------

Co-authored-by: Rootul Patel <rootulp@gmail.com>