docs: add minikube instructions for test/e2e (#3350)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

Adds simple instructions to run the test/e2e in minikube. 

Adding monitoring is not straightforward, as a lot of values are
hardcoded into the implementation. If there is a need to have the
monitoring locally, we can create a separate issue for it.

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>